[painter/gl] move #freeTexture() to common code
It no longer depends on the GL implementation.
