+++ internal/painter/gl/gl_core.go
@@ -11,7 +11,6 @@ import (
 	"github.com/go-gl/gl/v3.2-core/gl"
 
 	"fyne.io/fyne/v2"
-	"fyne.io/fyne/v2/internal/cache"
 	"fyne.io/fyne/v2/theme"
 )
 
@@ -34,17 +33,6 @@ type Buffer uint32
 type Program uint32
 
 var textureFilterToGL = []int32{gl.LINEAR, gl.NEAREST, gl.LINEAR}
-
-func (p *painter) freeTexture(obj fyne.CanvasObject) {
-	texture, ok := cache.GetTexture(obj)
-	if !ok {
-		return
-	}
-
-	p.ctx.DeleteTexture(Texture(texture))
-	p.logError()
-	cache.DeleteTexture(obj)
-}
 
 func (p *painter) glInit() {
 	err := gl.Init()
